requirements.txt
pandas
psycopg2~=2.9.10
requests~=2.32.3
sqlalchemy
python-dotenv~=1.1.0
bs4~=0.0.2
beautifulsoup4~=4.13.4